% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bsas}
\alias{bsas}
\title{bsas: Partidos de Buenos Aires}
\format{
\subsection{\code{bsas}}{

Este
\describe{
\item{partido}{Nombre del partido de prov de Buenos Aires}
\item{lat1}{Latitud}
\item{lat2}{Latitud}
\item{lng1}{Longitud}
\item{lng2}{Longitud}
\item{geom}{sfc_POLYGON}
}
}
}
\usage{
bsas
}
\description{
Dataset de clase sf con los polígonos de los partidos de Buenos Aires y
las coordenadas de la bounding box de cada uno. Ver el repositorio de
GeoCovid BsAs para mas detalle de estas Variables
}
\keyword{datasets}
