% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_conversion_fecha.R
\name{formatted_date}
\alias{formatted_date}
\title{Convierte el formato de la fecha de a "YYYY-MM-DD"}
\usage{
formatted_date(fecha)
}
\arguments{
\item{fecha}{Fecha}
}
\value{
Fecha en formato "YYYY-MM-DD"
}
\description{
Convierte el formato de la fecha de a "YYYY-MM-DD"
}
\examples{

formatted_date("")
 
}
