% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapa_covid_departamentos.R
\name{MapaCovidDepartamentos_UI}
\alias{MapaCovidDepartamentos_UI}
\title{IU: Mapa prov. de Buenos Aires por departamento}
\usage{
MapaCovidDepartamentos_UI(id, bsas)
}
\arguments{
\item{id}{}

\item{bsas}{Dataset de clase sf con los partidos de Buenos Aires.}
}
\value{
Elemntos de interfaz de usuario del tab 3
}
\description{
Este mapa se encuentra en el tab 3 de GeoCovid app.
}
