% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_data_sisa_deploy}
\alias{mini_data_sisa_deploy}
\title{mini_data_sisa_deploy: Extrato del dataset de casos de COVID-19}
\format{
\subsection{\code{mini_data_sisa_deploy}}{

Una extracto de la abse de datos con 2230 filas y 4 columnas:
\describe{
\item{id_evento_caso}{ID}
\item{residencia_provincia_nombre}{Provincia}
\item{residencia_departamento_nombre}{Departamento, partido}
\item{fecha_enfermo}{Fecha reporte caso COVID-19}
}
}
}
\usage{
mini_data_sisa_deploy
}
\description{
Ver el repositorio de GeoCovid BsAs para mas detalle de estas Variables
}
\keyword{datasets}
