% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{px_baires}
\alias{px_baires}
\title{px_baires: Valores promedio de movilidad ciudadana por partido}
\format{
\subsection{\code{px_baires}}{

Una extracto de la abse de datos con 4080 filas y 11 columnas:
\describe{
\item{fecha}{ID}
\item{locacion}{Provincia}
\item{tipo_de_raster}{Cambio porcentual prepandemia (pc) o semanal (7dpc)}
\item{partido}{Partido o departamento}
\item{noche_0}{Promedio para el raster de la noche}
\item{mañana_8}{Promedio para el raster de la mañana}
\item{tarde_16}{Promedio para el raster de la tarde}
\item{px_mean_dianoche}{Promedio para los rasters de mañana y tarde}
\item{criterio}{Vuelve variable categorica px_mean_dianoche}
\item{criterio_noche}{Vuelve variable categorica noche_0}
\item{geom}{sfc_POLYGON}
}
}
}
\usage{
px_baires
}
\description{
Ver el repositorio de GeoCovid BsAs para mas detalle de estas Variables.
Se esta empleando solo un extracto de la base de datos para este dataset de
ejempplo.
}
\keyword{datasets}
