% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{GeoCovid app Server}
\usage{
server(input, output, session, r)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{r}{}
}
\description{
GeoCovid app Server
}
