% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{GeoCovid app Server}
\usage{
server(input, output, session, r)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{r}{r}
}
\description{
GeoCovid app Server
}
